ref: move wizard-dialog to the module
