+++ src/wizarding/index.ts
@@ -2,3 +2,4 @@ export * from './wizard.js'
 export * from './events.js'
 export * from './dom.js'
 export * from './wizarding.js'
+export * from './wizard-dialog.js'
