+++ src/wizarding/wizarding.ts
@@ -9,7 +9,7 @@ import './wizard-dialog.js';
 // import { WizardDialog } from '../wizard-dialog.js';
 import { Wizard, WizardFactory } from './wizard.js';
 import { WizardEvent } from './events.js';
-import { WizardDialog } from '../wizard-dialog.js';
+import { WizardDialog } from './wizard-dialog.js';
 
 /** `LitElement` mixin that adds a `workflow` property which [[`Wizard`]]s are
  * queued onto on incoming [[`WizardEvent`]]s, first come first displayed. */
