+++ test/unit/wizard-dialog.test.ts
@@ -5,8 +5,8 @@ import './mock-editor.js';
 import { Button } from '@material/mwc-button';
 
 import '../../src/wizard-textfield.js';
-import '../../src/wizard-dialog.js';
-import { WizardDialog } from '../../src/wizard-dialog.js';
+import '../../src/wizarding/wizard-dialog.js';
+import { WizardDialog } from '../../src/wizarding/wizard-dialog.js';
 import { EditorAction, WizardInputElement } from '../../src/foundation.js';
 import { WizardCheckbox } from '../../src/wizard-checkbox.js';
 import { WizardSelect } from '../../src/wizard-select.js';
